requirements.txt
asgiref==3.4.1
certifi==2022.6.15
charset-normalizer==2.1.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==1.2.0
dj-static==0.0.6
Django==4.0.6
django-cors-headers==4.1.0
django-simple-history==3.1.1
djangorestframework==3.13.1
drf-yasg==1.21.3
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.32
packaging==21.3
Pillow==8.2.0
psycopg2==2.9.6
psycopg2-binary==2.9.5
pyparsing==3.0.9
python-dotenv==0.21.1
pytz==2021.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.1
static3==0.7.0
typing-extensions==3.10.0.0
tzdata==2022.1
Unipath==1.0
uritemplate==4.1.1
urllib3==1.26.11
vboxapi==1.0
whitenoise==6.3.0
